requirements.txt
# Core Dependencies
streamlit>=1.28.0
pandas>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0

# Data Visualization
altair>=5.0.0

# LangChain and RAG Dependencies
langchain>=0.1.0
langchain-openai>=0.0.8
langchain-community>=0.0.29
langchain-experimental>=0.0.57
langchain-core>=0.1.0

# Vector Database
chromadb>=0.4.0

# Geolocation
geoip2>=4.7.0

# HTTP Requests
requests>=2.31.0

# Standard Library (these come with Python)
# os, json, datetime, logging, glob, csv, typing, random